requirements.txt
aiohttp==3.8.1
aiopg==1.3.3
pyjwt==2.3.0
PyYAML==6.0
click==8.0.3
uvloop==0.16.0
peewee==3.14.8
peewee-async==0.7.2
marshmallow==3.14.1
aiojobs==1.0.0

# migrations
ex-peewee-moves==2.0.1

# scheduling
APScheduler==3.8.1

# pytest
pytest==6.2.5
pytest-aiohttp==0.3.0
aioresponses==0.7.3